fix(paywall): fail open, dismiss, drop bare-402 text match, app-level billing-updated listener
Review on #962:
- The card withheld the composer until a fresh snapshot said access.allowed
  === true, with no dismiss. Now it withholds only while a fresh snapshot
  positively denies (allowed === false); in every other state (no access
  field, stale fallback, fetch error, platform disconnected, loading) the
  card renders above the composer. Dismiss always hands the composer back.
- isInsufficientBalance no longer matches a bare 402 in the text; a
  parenthesized (500) with '402-abc' elsewhere stays null.
- Dead CTA computed during isLoading removed; loading shows only the spinner.
- billing-updated deep link: one BillingUpdatedHandler in RootLayout
  invalidates ['platform-billing']. pendingBillingUpdated flag, the
  flush-pending-billing-updated handle, billing-updated-consumed and the
  preload finally are gone.
- resolveOrgLink(path, org) in error-presentation.ts is the single {orgId}
  resolver (used by resolvePresentationMarkdown and the paywall CTA);
  ORG_BILLING_LINK is exported instead of re-declared; the two identical
  status sets are one.
- paywall-card.tsx split into paywall-cta.ts (pure, node tests),
  use-paywall-billing.ts and paywall-card.tsx.
